+++ test/dummy/app/components/dashboard/application_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Dashboard
+  class ApplicationComponent < ::Vident::Phlex::HTML
+    include ::Phlex::Rails::Helpers::Routes
+  end
+end