+++ test/dummy/config/initializers/vident.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+Vident::StableId.strategy = if Rails.env.test?
+  Vident::StableId::RANDOM_FALLBACK
+else
+  Vident::StableId::STRICT
+end